requirements.txt
aiokafka==0.10.0
annotated-types==0.7.0
async-timeout==4.0.3
dnspython==2.6.1
motor==3.4.0
packaging==24.1
pydantic==2.7.4
pydantic_core==2.18.4
pymongo==4.7.3
python-dotenv==1.0.1
typing_extensions==4.12.2
